% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recursive_col.R
\name{E_soft_col}
\alias{E_soft_col}
\title{Softmax Expected Value and Optimal Next Decision in the Lose Condition}
\usage{
E_soft_col(
  e,
  n_trials,
  n_players,
  r1,
  r2,
  r3,
  s1,
  s2,
  payoff_r2 = 0,
  payoff_r3 = 10,
  payoff_s1 = 0,
  payoff_s2 = 1,
  coord = FALSE,
  inv_temp = 10
)
}
\arguments{
\item{e}{The current endowment}

\item{n_trials}{The remaining number of trials for all players (e.g., if 5 players play 5 trials each this number would be 25)}

\item{n_players}{The number of players}

\item{r1}{The probability of extinction when choosing the risky lottery (outcome 1)}

\item{r2}{The probability of outcome 2 when choosing the risky lottery}

\item{r3}{The probability of outcome 3 when choosing the risky lottery}

\item{s1}{the probability of outcome 1 when choosing the safe lottery}

\item{s2}{the probability of outcome 2 when choosing the safe lottery}

\item{payoff_r2}{The payoff of outcome 2 when choosing the risky lottery}

\item{payoff_r3}{The payoff of outcome 3 when choosing the risky lottery}

\item{payoff_s1}{The payoff of outcome 1 when choosing the safe lottery}

\item{payoff_s2}{The payoff of outcome 2 when choosing the safe lottery}

\item{coord}{Whether players can coordinate (e.g., through text) to hit a specific target number of risky choices or not (beta)}

\item{inv_temp}{Softmax inverse temperature}
}
\value{
A list with the expected payoff when following the optimal strategy and the optimal next choice.
}
\description{
This function calculates the optimal choice and associated expected payoff given the endowment, the remaining trials, and the
payoffs and associated probabilities for the Lose Condition (i.e., condition, where participants don't get any bonus if they draw the extinction event).
}
\examples{
cache_soft_col <- list()
E_soft_col(e = 0, n_trials = 20, n_players = 5, r1 = 0.05, r2 = 0.475, 
r3 = 0.475, s1 = 0.5, s2 = 0.5, payoff_r2 = 0, payoff_r3 = 10, payoff_s1 = 0, 
payoff_s2 = 1)
}
