% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recursive_losecontinue.R
\name{E_soft_reset}
\alias{E_soft_reset}
\title{Expected Value and Probability of Next Decision of Optimal Strategy in the Lose Continue Condition Using Softmax}
\usage{
E_soft_reset(
  e,
  n,
  r1,
  r2,
  r3,
  s1,
  s2,
  inv_temp = 10,
  payoff_r2 = 0,
  payoff_r3 = 10,
  payoff_s1 = 0,
  payoff_s2 = 1
)
}
\arguments{
\item{e}{The current endowment}

\item{n}{The remaining number of trials}

\item{r1}{The probability of extinction when choosing the risky lottery (outcome 1)}

\item{r2}{The probability of outcome 2 when choosing the risky lottery}

\item{r3}{The probability of outcome 3 when choosing the risky lottery}

\item{s1}{the probability of outcome 1 when choosing the safe lottery}

\item{s2}{the probability of outcome 2 when choosing the safe lottery}

\item{inv_temp}{Inverse temperature for the softmax decision function}

\item{payoff_r2}{The payoff of outcome 2 when choosing the risky lottery}

\item{payoff_r3}{The payoff of outcome 3 when choosing the risky lottery}

\item{payoff_s1}{The payoff of outcome 1 when choosing the safe lottery}

\item{payoff_s2}{The payoff of outcome 2 when choosing the safe lottery}
}
\value{
A list with the expected payoff when following the optimal softmax strategy and the probability of playing safe and risky
}
\description{
This function calculates the optimal choice and associated expected payoff given the endowment, the remaining trials, and the
payoffs and associated probabilities for the Lose Condition
(i.e., condition, where participants lose everything but can keep playing).
Unlike E_reset it uses a softmax transformation to convert the difference
in EV between the risky and safe choice to a choice probability
}
\examples{
soft_cache_lc <- list()
E_soft_reset(e = 0, n = 100, r1 = 0.05, r2 = 0.475, r3 = 0.475, s1 = 0.5, 
s2 = 0.5, payoff_r2 = 0, payoff_r3 = 10, payoff_s1 = 0, payoff_s2 = 1)
}
