% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EV_keep.R
\name{max_ev_keep}
\alias{max_ev_keep}
\title{Optimal Number of Risky Choices in the Keep Condition}
\usage{
max_ev_keep(
  n,
  e,
  r1,
  r2,
  r3,
  s1,
  s2,
  payoff_r2 = 0,
  payoff_r3 = 10,
  payoff_s1 = 0,
  payoff_s2 = 1
)
}
\arguments{
\item{n}{The remaining number of trials}

\item{e}{The endowment}

\item{r1}{The probability of extinction when choosing the risky lottery (outcome 1)}

\item{r2}{The probability of outcome 2 when choosing the risky lottery}

\item{r3}{The probability of outcome 3 when choosing the risky lottery}

\item{s1}{the probability of outcome 1 when choosing the safe lottery}

\item{s2}{the probability of outcome 2 when choosing the safe lottery}

\item{payoff_r2}{The payoff of outcome 2 when choosing the risky lottery}

\item{payoff_r3}{The payoff of outcome 3 when choosing the risky lottery}

\item{payoff_s1}{The payoff of outcome 1 when choosing the safe lottery}

\item{payoff_s2}{The payoff of outcome 2 when choosing the safe lottery}
}
\value{
A list with the expected payoff when following the optimal strategy, the optimal next choice, a
and a softmax transformation with inverse temperature 50 containing the probabilities of the next choice
(this helps pick up on cases, where there are only tiny EV differences between playing safe and risky)
}
\description{
Loops over ev_keep to estimate the optimal number of risky choices
}
\examples{
#Optimal Strategy for Maier et al. (2024)
max_ev_keep(e = 0, n = 100, r1 = 0.05, r2 = 0.475, r3 = 0.475, s1 = 0.5,
s2 = 0.5, payoff_r2 = 0, payoff_r3 = 10, payoff_s1 = 0, payoff_s2 = 1)
}
