% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGambling.R
\docType{package}
\name{package-package}
\alias{XGambling}
\alias{XGambling-package}
\alias{package-package}
\title{XGambling}
\description{
Functions for the Extinction Gambling Task
}
\details{
Click on 'index' below to see all the functions.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
