% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Function to Simulate a Large Number of Agents Playing the Task and Following the Optimal Strategy}
\usage{
simulate(
  e,
  max_n,
  r1,
  r2,
  r3,
  s1,
  s2,
  payoff_r2 = 0,
  payoff_r3 = 10,
  payoff_s1 = 0,
  payoff_s2 = 1,
  soft = TRUE,
  type = "lose",
  inv_temp = 10,
  precision = 1000,
  amb_risk = 0
)
}
\arguments{
\item{e}{The current endowment}

\item{max_n}{The remaining number of trials}

\item{r1}{The probability of extinction when choosing the risky lottery (outcome 1)}

\item{r2}{The probability of outcome 2 when choosing the risky lottery}

\item{r3}{The probability of outcome 3 when choosing the risky lottery}

\item{s1}{the probability of outcome 1 when choosing the safe lottery}

\item{s2}{the probability of outcome 2 when choosing the safe lottery}

\item{payoff_r2}{The payoff of outcome 2 when choosing the risky lottery}

\item{payoff_r3}{The payoff of outcome 3 when choosing the risky lottery}

\item{payoff_s1}{The payoff of outcome 1 when choosing the safe lottery}

\item{payoff_s2}{The payoff of outcome 2 when choosing the safe lottery}

\item{soft}{whether softmax or deterministic version should be used (TRUE or FALSE)}

\item{type}{whether "lose" or "reset" version is used}

\item{inv_temp}{inverse temperature for the softmax transformation}

\item{precision}{the number of agents being simulated}

\item{amb_risk}{The background extinction risk independent of the players action (currently only for E_lose).}
}
\value{
A list containing "total_winnings" a vector with the earnings for each agent, "num_extinctions" containing the number of extinctions,
"all_risky_plays" a dataframe with the risky plays over time for each agent, and "all_alive" denoting for each agent and trial whether they are alive.
}
\description{
This function simulates a number of agents who play the extinction gambling task following the optimal dynamic strategy.
It only applies to the Lose and the Reset condition as the Keep condition does not have a dynamic strategy
}
\examples{
soft_cache <- list()
simulate(e = 0, max_n = 10, r1 = 0.05, r2 = 0.475, r3 = 0.475, s1 = 0.5, 
s2 = 0.5, payoff_r2 = 0, payoff_r3 = 10, payoff_s1 = 0, payoff_s2 = 1, 
soft = TRUE, type = "lose", inv_temp = 10, precision = 10)
}
