% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recursive_col.R, R/sim_play_col.R
\name{play_col}
\alias{play_col}
\title{Simulate a Single Round of the Lose Condition with Coordination Options}
\usage{
play_col(
  e,
  max_n,
  n_players,
  r1,
  r2,
  r3,
  s1,
  s2,
  payoff_r2 = 0,
  payoff_r3 = 10,
  payoff_s1 = 0,
  payoff_s2 = 1,
  coord = FALSE
)

play_col(
  e,
  max_n,
  n_players,
  r1,
  r2,
  r3,
  s1,
  s2,
  payoff_r2 = 0,
  payoff_r3 = 10,
  payoff_s1 = 0,
  payoff_s2 = 1,
  coord = FALSE
)
}
\arguments{
\item{e}{The current endowment}

\item{max_n}{The remaining number of trials for all players (e.g., if 5 players play 5 trials each this number would be 25)}

\item{n_players}{The number of players}

\item{r1}{The probability of extinction when choosing the risky lottery (outcome 1)}

\item{r2}{The probability of outcome 2 when choosing the risky lottery}

\item{r3}{The probability of outcome 3 when choosing the risky lottery}

\item{s1}{the probability of outcome 1 when choosing the safe lottery}

\item{s2}{the probability of outcome 2 when choosing the safe lottery}

\item{payoff_r2}{The payoff of outcome 2 when choosing the risky lottery}

\item{payoff_r3}{The payoff of outcome 3 when choosing the risky lottery}

\item{payoff_s1}{The payoff of outcome 1 when choosing the safe lottery}

\item{payoff_s2}{The payoff of outcome 2 when choosing the safe lottery}

\item{coord}{Whether players can coordinate (e.g., through text) to hit a specific target number of risky choices or not}

\item{soft}{Should softmax version be used?}

\item{inv_temp}{Inverse temperature for softmax}
}
\value{
A list containing the following elements:
\item{cumulative_winnings}{A vector of the cumulative winnings over the course of the round}
\item{risky_plays}{A logical vector indicating whether risky plays were made in each round}
\item{rounds}{The total number of risky plays made}
\item{extinction}{A boolean indicating whether an extinction event occurred during the round}
\item{rounds_alive}{A logical vector indicating whether players remained alive during each round of the game}

A list with the expected payoff when following the optimal strategy and the optimal next choice.
}
\description{
This function simulates a single round of the Lose Condition (where players do not get any bonus if they draw the extinction event).
It tracks cumulative winnings, the number of risky plays, the number of rounds completed, and whether extinction occurred.

This function simulates n agents that plays the extinction gambling task following the optimal dynamic strategy.
}
\examples{
# Example: Simulate a single round with default payoffs and no coordination
cache_col <- list()
play_col(e = 0, max_n = 20, n_players = 5, r1 = 0.05, r2 = 0.475, r3 = 0.475, 
         s1 = 0.5, s2 = 0.5)
cache_col <- list()
play_col(e = 0, max_n = 20, n_players = 5, r1 = 0.05, r2 = 0.475, 
r3 = 0.475, s1 = 0.5, s2 = 0.5, payoff_r2 = 0, payoff_r3 = 10, payoff_s1 = 0, 
payoff_s2 = 1, coord = FALSE)
}
