% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_play_col.R
\name{simulate_col}
\alias{simulate_col}
\title{Function to Simulate Many Runds of the Collective Goods Game Version of the Extinction Gambling Task}
\usage{
simulate_col(
  e,
  max_n,
  n_players,
  r1,
  r2,
  r3,
  s1,
  s2,
  payoff_r2 = 0,
  payoff_r3 = 10,
  payoff_s1 = 0,
  payoff_s2 = 1,
  coord = FALSE,
  precision = 1000
)
}
\arguments{
\item{e}{The current endowment}

\item{max_n}{The remaining number of trials for all players (e.g., if 5 players play 5 trials each this number would be 25)}

\item{n_players}{The number of players}

\item{r1}{The probability of extinction when choosing the risky lottery (outcome 1)}

\item{r2}{The probability of outcome 2 when choosing the risky lottery}

\item{r3}{The probability of outcome 3 when choosing the risky lottery}

\item{s1}{the probability of outcome 1 when choosing the safe lottery}

\item{s2}{the probability of outcome 2 when choosing the safe lottery}

\item{payoff_r2}{The payoff of outcome 2 when choosing the risky lottery}

\item{payoff_r3}{The payoff of outcome 3 when choosing the risky lottery}

\item{payoff_s1}{The payoff of outcome 1 when choosing the safe lottery}

\item{payoff_s2}{The payoff of outcome 2 when choosing the safe lottery}

\item{coord}{Whether players can coordinate (e.g., through text) to hit a specific target number of risky choices or not}

\item{precision}{How many games should be simulated}
}
\value{
A list with the expected payoff when following the optimal strategy and the optimal next choice.
}
\description{
This function simulates a number of games of players playing the collective extintion gambling task and following the optimal dynamic strategy.
}
\examples{
cache_col <- list()
simulate_col(e = 0, max_n = 20, n_players = 5, r1 = 0.05, r2 = 0.475, 
r3 = 0.475, s1 = 0.5, s2 = 0.5, payoff_r2 = 0, payoff_r3 = 10, payoff_s1 = 0, 
payoff_s2 = 1, coord = FALSE, precision = 1000)
}
