% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_play_col.R
\name{simulate_col}
\alias{simulate_col}
\title{Simulate Multiple Playthroughs of the Lose Condition with Coordination Options}
\usage{
simulate_col(
  e,
  max_n,
  n_players,
  r1,
  r2,
  r3,
  s1,
  s2,
  payoff_r2 = 0,
  payoff_r3 = 10,
  payoff_s1 = 0,
  payoff_s2 = 1,
  precision = 1000,
  clearCache = "None",
  coord = FALSE,
  soft = FALSE,
  inv_temp = 10
)
}
\arguments{
\item{e}{The starting endowment (initial winnings) for each player}

\item{max_n}{The maximum number of trials (total rounds) for the simulation}

\item{n_players}{The number of players participating in the simulation}

\item{r1}{The probability of extinction when choosing the risky lottery (outcome 1)}

\item{r2}{The probability of outcome 2 when choosing the risky lottery}

\item{r3}{The probability of outcome 3 when choosing the risky lottery}

\item{s1}{The probability of outcome 1 when choosing the safe lottery}

\item{s2}{The probability of outcome 2 when choosing the safe lottery}

\item{payoff_r2}{The payoff of outcome 2 when choosing the risky lottery (default = 0)}

\item{payoff_r3}{The payoff of outcome 3 when choosing the risky lottery (default = 10)}

\item{payoff_s1}{The payoff of outcome 1 when choosing the safe lottery (default = 0)}

\item{payoff_s2}{The payoff of outcome 2 when choosing the safe lottery (default = 1)}

\item{precision}{The number of iterations to simulate (default = 1000)}

\item{clearCache}{Set to "col" to clear the cache}

\item{coord}{Boolean indicating whether players can coordinate on risky choices or not (default = FALSE)}

\item{soft}{Should softmax version be used?}

\item{inv_temp}{Inverse temperature for softmax}
}
\value{
A list containing the following elements:
\item{outcome}{The outcome of the last round of simulation (includes cumulative winnings, risky plays, etc.)}
\item{total_winnings}{A vector of total winnings for each iteration}
\item{num_extinctions}{The total number of extinctions that occurred across all iterations}
\item{all_risky_plays}{A matrix tracking the risky plays for each player across all iterations}
\item{all_alive}{A matrix tracking whether players remained alive across all iterations}
}
\description{
This function simulates multiple rounds of the Lose Condition (where players do not get any bonus if they draw the extinction event).
It tracks cumulative winnings, the number of extinctions, the number of risky plays, and rounds alive across multiple iterations.
}
\examples{
cache_col <- list()
simulate_col(e = 0, max_n = 20, n_players = 5, r1 = 0.05, r2 = 0.475, r3 = 0.475, 
             s1 = 0.5, s2 = 0.5, payoff_r2 = 5, payoff_r3 = 20, 
             payoff_s1 = 0, payoff_s2 = 2, precision = 1000, coord = TRUE)
}
